+++ pool/poolbuilder.cpp
@@ -1,12 +1,12 @@
 #include "poolbuilder.h"
 #include "poolfactory.h"
 
-Ball* PoolBuilder::buildBall(Coordinate coordinate, std::string color, int mass, int radius, double xVelocity, double yVelocity) {
+Ball* PoolBuilder::buildBall(Coordinate coordinate, std::string color, float mass, float radius, float xVelocity, float yVelocity) {
     PoolFactory factory;
     return factory.createBall(coordinate, color, mass, radius, xVelocity, yVelocity);
 }
 
-Table* PoolBuilder::buildTable(unsigned int length, unsigned int height, unsigned int thickness, unsigned int space, unsigned int friction) {
+Table* PoolBuilder::buildTable(float length, float height, float thickness, float space, float friction) {
     PoolFactory factory;
     return factory.createTable(length, height, thickness, space, friction);
 }
